requirements.txt
flask
pandas
numpy
plotly
openpyxl
gunicorn
sqlalchemy
psycopg2-binary
